support only DQA in IQMBackendBase
