+++ src/iqm/qiskit_iqm/iqm_backend.py
@@ -17,54 +17,17 @@ from __future__ import annotations
 
 from abc import ABC
 import itertools
-from typing import Final, Optional, Union
-from uuid import UUID
+from typing import Final, Optional
 
 from qiskit.circuit import Parameter
 from qiskit.circuit.library import CZGate, IGate, Measure, RGate
 from qiskit.providers import BackendV2
 from qiskit.transpiler import InstructionProperties, Target
 
-from iqm.iqm_client import (
-    DynamicQuantumArchitecture,
-    GateImplementationInfo,
-    GateInfo,
-    QuantumArchitectureSpecification,
-)
+from iqm.iqm_client import DynamicQuantumArchitecture
 from iqm.qiskit_iqm.move_gate import MoveGate
 
 IQM_TO_QISKIT_GATE_NAME: Final[dict[str, str]] = {'prx': 'r', 'cz': 'cz'}
-
-
-def _dqa_from_static_architecture(sqa: QuantumArchitectureSpecification) -> DynamicQuantumArchitecture:
-    """Create a dynamic quantum architecture from the given static quantum architecture.
-
-    Since the DQA contains some attributes that are not present in an SQA, they are filled with mock data:
-
-    * Each gate type is given a single mock implementation.
-    * Calibration set ID is set to the all-zeros UUID.
-
-    Args:
-        sqa: static quantum architecture to replicate
-    Returns:
-        DQA replicating the properties of ``sqa``
-    """
-    qubits = [qb for qb in sqa.qubits if qb.startswith('QB')]
-    computational_resonators = [qb for qb in sqa.qubits if qb.startswith('COMP')]
-    gates = {
-        gate_name: GateInfo(
-            implementations={'__fake': GateImplementationInfo(loci=tuple(tuple(locus) for locus in gate_loci))},
-            default_implementation='__fake',
-            override_default_implementation={},
-        )
-        for gate_name, gate_loci in sqa.operations.items()
-    }
-    return DynamicQuantumArchitecture(
-        calibration_set_id=UUID('00000000-0000-0000-0000-000000000000'),
-        qubits=qubits,
-        computational_resonators=computational_resonators,
-        gates=gates,
-    )
 
 
 class IQMBackendBase(BackendV2, ABC):
@@ -78,18 +41,13 @@ class IQMBackendBase(BackendV2, ABC):
 
     def __init__(
         self,
-        architecture: Union[QuantumArchitectureSpecification, DynamicQuantumArchitecture],
+        architecture: DynamicQuantumArchitecture,
         **kwargs,
     ):
         super().__init__(**kwargs)
-        if isinstance(architecture, QuantumArchitectureSpecification):
-            arch = _dqa_from_static_architecture(architecture)
-        else:
-            arch = architecture
-        self.architecture = arch
 
-        qb_to_idx = {qb: idx for idx, qb in enumerate(arch.components)}
-        operations = {gate_name: gate_info.loci for gate_name, gate_info in arch.gates.items()}
+        qb_to_idx = {qb: idx for idx, qb in enumerate(architecture.components)}
+        operations = {gate_name: gate_info.loci for gate_name, gate_info in architecture.gates.items()}
         target = Target()
 
         def _create_properties(
@@ -110,7 +68,7 @@ class IQMBackendBase(BackendV2, ABC):
             target.add_instruction(Measure(), _create_properties('measure'))
         target.add_instruction(
             IGate(),
-            {(qb_to_idx[qb],): None for qb in arch.components},
+            {(qb_to_idx[qb],): None for qb in architecture.components},
         )
         if 'prx' in operations:
             target.add_instruction(RGate(Parameter('theta'), Parameter('phi')), _create_properties('prx'))
@@ -123,6 +81,7 @@ class IQMBackendBase(BackendV2, ABC):
         self._qb_to_idx = qb_to_idx
         self._idx_to_qb = {v: k for k, v in qb_to_idx.items()}
         self.name = 'IQMBackend'
+        self.architecture = architecture
 
     @property
     def target(self) -> Target:
